Added validation on method Set
It's not possible to set a value when the database is in read-only mode.
This feature is not ready yet: despite the mode in Go code, the kcdbopen
function is always called with KCOWRITER|KCOCREATE mode, which indicates
read and write mode.

When the client is in read-only mode, the kcdbopen function must be
invoked with the KCOREADER constant.